feature(targeted-methods): creates the C library for getting the energy footprint of a piece of code in a C program.
Creates the library with the three directives that can be used to open the needed counters for a piece of code and to send the associated file descriptors to PowerAPI, to feed the counters with the following instructions and to stop feeding them.
PowerAPI is then responsible to collect periodically the opened file descriptors for producing the energy footprint of a piece of code.